+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2019-07-30
+### Changed
+- Update pre-commit to version 1.17.0
+- Update huey to version 2.1.0
+- Update redis to version 3.3.0
+- Use pytest-flask instead of Flask-Testing
+- Use pytest instead of unittest
+
 ## [1.0.2] - 2019-05-14
 ### Fixed
 - [Huey] Redis dependency was missing.